% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HurdleLikelihoodClass.R
\docType{methods}
\name{$,HurdleLikelihood-method}
\alias{$,HurdleLikelihood-method}
\title{Internal functions}
\usage{
\S4method{$}{HurdleLikelihood}(x, name = "LLall")
}
\arguments{
\item{x}{\code{HurdleLikelihood} class}

\item{name}{C++ method to call}
}
\value{
depends
}
\description{
Man page for internal functions
}
