% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GroupLasso.R
\name{cgpaths}
\alias{cgpaths}
\title{Get a solution path for the CG model}
\usage{
cgpaths(y.zif, this.model, nlambda = 100, lambda, lambda.min.ratio = if
  (length(y.zif) < ncol(this.model)) 0.005 else 0.05, control = list(tol =
  1e-05, maxrounds = 300, maxit = 500, debug = 1), standardize = FALSE)
}
\arguments{
\item{y.zif}{(zero-inflated) response}

\item{this.model}{(zero-inflated) covariates}

\item{lambda}{penalty path desired}

\item{control}{optimization control parameters}

\item{theta}{initial guess for parameter}
}
\value{
matrix of parameters, one row per lambda
}
\description{
Get a solution path for the CG model
}

