% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NeighborhoodEstimation.R
\name{collectStability}
\alias{collectStability}
\title{Process stability selected networks}
\usage{
collectStability(stabout, theta = 0.1, stabilityCheckpointDir = NULL)
}
\arguments{
\item{stabout}{output of \code{\link{stability}}}

\item{theta}{sparsity of the solution in terms of the number of parameters. default .1.}

\item{stabilityCheckpointDir}{path to a directory to write checkpoint files}
}
\value{
A list with components.  You probably want component 3, giving the stability coefficients for each edge.
}
\description{
Stability selection attempts to estimate the probability
that an edge would have been included when we sample from the population.
Following Shah and Samsworth, we process the stability selection to give stability coefficients for each edge.
From Shah and Samsworth: p = #edges = m^2; q = max(iknot) and theta = q/p.
We bound the set of population transient edges, which have probabilities of selection below
theta (the same theta as above for convenience) when sampling from the population.
We call an edge \emph{empirically transient} when the stability coefficient is below tau.
Then for a sparsity theta < .1 and stability coefficient tau > .7, the ratio of empirical transient edges to population edges is less than 1%, and in fact typically more like .1%.
(Table 2 of Shah and Samworth)
}
