% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NeighborhoodEstimation.R
\name{conditionalCenter}
\alias{conditionalCenter}
\title{Center data so that it has conditional column mean zero}
\usage{
conditionalCenter(samp)
}
\arguments{
\item{samp}{matrix of data}
}
\value{
centered matrix
}
\description{
Center data so that it has conditional column mean zero
}
\examples{
dat = data.frame(x=c(1:5, 0, 0, 0), y=c(6:10, 0, 0, 0))
conditionalCenter(dat)
#compare to
scale(dat, scale=FALSE)
}
