% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HurdleDistributions.R
\name{.evalGridHurdleMeasure}
\alias{.evalGridHurdleMeasure}
\title{Evaluate a Riemann sum on a grid of points that includes zero}
\usage{
.evalGridHurdleMeasure(lower, upper, dx, fun, ..., tol = 0.05)
}
\arguments{
\item{lower}{vector of lower limits}

\item{upper}{upper limits}

\item{dx}{scalar gridsize}

\item{fun}{function to evaluate.  First argument should be the vector point at which it's being eval'd}

\item{...}{passed to fun}

\item{tol}{passed to fun}
}
\value{
list of: reimann-weighted function on the grid/lattice (feval) and
breaks (lattice points)
}
\description{
Evaluate a Riemann sum on a grid of points that includes zero
}

