% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Geom3.R, R/_Geom3_R
\name{geom_hurdle}
\alias{geom_hurdle}
\title{Plot conditional regression line and conditional means}
\usage{
geom_hurdle(mapping = NULL, data = NULL, stat = "identity",
  position = "identity", na.rm = FALSE, show.legent = NA,
  inherit.aes = TRUE, anno.colors = c("blue", "red"), tol = 0.01, ...)

geom_hurdle(mapping = NULL, data = NULL, stat = "identity",
  position = "identity", na.rm = FALSE, show.legent = NA,
  inherit.aes = TRUE, anno.colors = c("blue", "red"), tol = 0.01, ...)
}
\arguments{
\item{mapping}{}

\item{data}{}

\item{stat}{}

\item{position}{}

\item{na.rm}{}

\item{anno.colors}{}

\item{tol}{}

\item{...}{}

\item{mapping}{}

\item{data}{}

\item{stat}{}

\item{position}{}

\item{na.rm}{}

\item{anno.colors}{}

\item{tol}{}

\item{...}{}
}
\value{
plot

plot
}
\description{
Details

Details
}

