% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/HurdleStructure.R
\name{getConditionalMLE}
\alias{getConditionalMLE}
\title{Loop over indices and solve condiiton MLE}
\usage{
getConditionalMLE(hs, using = "gibbs", testGrad = FALSE, engine = "R",
  ...)
}
\arguments{
\item{hs}{HurdleStructure}

\item{using}{'gibbs' or 'sample'}

\item{...}{passed to optim}
}
\value{
list of length two giving parameter values and standard errors. j is the index of the response, i is the index of the coefficient.
}
\description{
Loop over indices and solve condiiton MLE
}

