% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HurdleStructure.R
\name{getGibbs}
\alias{getGibbs}
\title{Update the current gibbs sample for a \code{HurdleStructure}}
\usage{
getGibbs(hs, Nkeep = 2000, kcell, post_fun, mvnorm, ...)
}
\arguments{
\item{hs}{\code{HurdleStructure}}

\item{Nkeep}{final number of samples to keep}

\item{kcell}{Number of cells per sample (cells added on exponential scale)}

\item{post_fun}{Function to apply post-sampling, ie, to add contaminating noise}

\item{mvnorm}{Sample from multivariate normal (using only K matrix)--intended for applying post-selection models}

\item{...}{passed to \code{rGibbsHurdle}}
}
\value{
modified \code{HurdleStructure}
}
\description{
Replaces the item \code{'gibbs'}
}
