% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProcessNeighborhoods.R
\name{getSafeApprox}
\alias{getSafeApprox}
\title{Return a function interpolating over a grid}
\usage{
getSafeApprox(lpath)
}
\arguments{
\item{lpath}{grid of points over which interpolation is required}
}
\value{
function of (x,y) providing the interpolations for x=lpath
}
\description{
In the interior of x, interpolation is used, otherwise the value of y at the endpoint is used.
Returns an empty list if length(x) = 0.
}
