% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotNetworks.R
\name{hn_plot}
\alias{hn_plot}
\title{Plot an igraph object}
\usage{
hn_plot(gadj, layout = layout_with_kk, width.scale = 2, colorbar = FALSE,
  colorbarx, colorbary, ...)
}
\arguments{
\item{gadj}{igraph object}

\item{layout}{layout function}

\item{width.scale}{scale factor for edges}

\item{colorbar}{\code{logical}, should a legend providing a scale for the edges be displayed?}

\item{colorbarx}{coordinates for legend}

\item{colorbary}{coordinates for legend}

\item{...}{passed to plot.igraph}
}
\value{
plots the network
}
\description{
Plot an igraph object
}
