% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HurdleNormal-package.R
\docType{package}
\name{HurdleNormal-package}
\alias{HurdleNormal}
\alias{HurdleNormal-package}
\title{A Hurdle model is a modification of a distribution at zero. This
   package provides routines to estimate and sample from multivariate Hurdle models
   on a Normal density. These are distributions that are conditionally Normal, but with singularities along the coordinate axes, so generalize a univariate zero-inflated distribution.
The main functionality is to estimate graphical models using group-lasso penalized neighborhood selection.}
\description{
A Hurdle model is a modification of a distribution at zero. This package provides routines to estimate and sample from multivariate Hurdle models on a Normal density. These are distributions that are conditionally Normal, but with singularities along the coordinate axes, so generalize a univariate zero-inflated distribution. The main functionality is to estimate graphical models using group-lasso penalized neighborhood selection, and has applications to modeling gene-gene interactions in single cell gene expression.
}
\references{
https://arxiv.org/abs/1610.05857
}
\author{
\strong{Maintainer}: Andrew McDavid \email{Andrew_McDavid@urmc.rochester.edu}

}
