% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HurdleStructure.R
\name{HurdleStructure}
\alias{HurdleStructure}
\title{Structure to hold a model, samples and estimates}
\usage{
HurdleStructure(G = NULL, H = NULL, K = NULL, sample = NULL,
  gibbs = TRUE, estimate = NULL, gibbs_args = list(kcell = 1, mvnorm =
  FALSE))
}
\arguments{
\item{G}{symmetric discrete interaction matrix}

\item{H}{unstructured location matrix}

\item{K}{Symmetric positive definite conditional precision matrix}

\item{sample}{optional sample from the model}

\item{gibbs}{If \code{TRUE}, then sample from the provided model.  Else may contain a sample from the model}

\item{estimate}{optional estimated coefficients (replaces provided \code{G}, \code{H}, \code{K} if provided).}

\item{gibbs_args}{list of arguments passed to \code{\link{getGibbs}}}
}
\value{
\code{list} with the above components
}
\description{
Structure to hold a model, samples and estimates
}
\author{
Andrew McDavid
}
