% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GroupLasso.R
\name{makeModelOrthog}
\alias{makeModelOrthog}
\title{Generate a design matrix corresponding to the 2-1-0 hurdle}
\usage{
makeModelOrthog(zif, ..., tol = 0.001)
}
\arguments{
\item{zif}{zero-inflated covariates}

\item{...}{arguments passed to `scale`}

\item{tol}{tolerance for declaring a parameter equal to zero}
}
\value{
design matrix
}
\description{
details
}

