% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AutoLogistic.R
\name{neighborhoodToArray}
\alias{neighborhoodToArray}
\title{Convert neighborhood estimates fit on differing lambda paths into adjacency matrices}
\usage{
neighborhoodToArray(pathList, Coordmap, vnames = NULL)
}
\arguments{
\item{pathList}{a list of paths (coefficients in columns, paths in rows)}

\item{Coordmap}{A mapping (vector of length ncol(pathList[[1]])) between coefficient-indices and variables.  See details.}

\item{vnames}{a vector of names to be applied to the resulting adjacency matrix}
}
\value{
list of (sparse) adjacency matrices, the number of non-zero elements for each matrix, and the lambda for each matrix.
}
\description{
Join a series of paths fit node-wise into a list of adjacency matrices.
}
\details{
The \code{Coordmap} can be missing when \code{cgpaths} has been run, as the mapping is provided in the output.  Otherwise, \code{Coordmap} should be a vector, giving the node index for each coefficient in the model (to accodomdate models with several coefficients per node).  The node index `1` is assumed to correspond to intercept (nuisance) terms.
}

