% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Util.R
\name{promoteWarning}
\alias{promoteWarning}
\title{Selectively convert a warnings to an error}
\usage{
promoteWarning(expr, regexp)
}
\arguments{
\item{expr}{an expression}

\item{regexp}{a regexp to be matched (with str_detect)}
}
\value{
the result of expr
}
\description{
Selectively convert a warnings to an error
}
