% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HurdleDistributions.R
\name{rCondHurdle210}
\alias{rCondHurdle210}
\title{Sample from the full conditional distribution}
\usage{
rCondHurdle210(x, j, G, H, K, tol = 5e-04)
}
\arguments{
\item{x}{vector, or matrix, to be conditioned upon.}

\item{j}{coordinate to be returned}

\item{G}{binary dependence matrix}

\item{H}{location dependence matrix}

\item{K}{precision matrix}

\item{tol}{tolerance for calling a point to be numerically equal to zero.}
}
\value{
numeric sample from x_i | x_{-i}
}
\description{
Return a sample from f(x_i | x_{-i}), that is, a sample of the conditional distribution of the \code{i}th coordinates, given \code{x} aside from \code{x_i}.
}
\details{
\code{x} is assumed in the same order as the parameter vectors, aside from its missing ith element.
This function is vectorized over x.
}

