% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HurdleDistributions.R
\name{rGibbsHurdle}
\alias{rGibbsHurdle}
\alias{cpp_rGibbsHurdle}
\title{Sample from a multivariate hurdle model}
\usage{
rGibbsHurdle(G, H, K, Nt, burnin = 500, thin = 0.1, tol = 5e-04, Nkeep = 500)
}
\arguments{
\item{G}{symmetric discrete interaction matrix}

\item{H}{unstructured location matrix}

\item{K}{Symmetric positive definite conditional precision matrix}

\item{Nt}{Number of unthinned samples, including burnin}

\item{burnin}{how many samples to discard from burn in}

\item{thin}{how many samples to thin}

\item{tol}{Numeric tolerance for zero}

\item{Nkeep}{(optional) number of samples, post-burnin and thinning}
}
\value{
matrix of (Nt-burnin)*thin samples
}
\description{
Sample from a multivariate hurdle model
}
\examples{
G = matrix(c(-15, 1, 0,
1, -15, 1.5,
0, 1.5, -15), nrow=3)
H = diag(5, nrow=3)
K = diag(1, nrow=3)
y = rGibbsHurdle(G, H, K, 2000, thin = .2, burnin=1000)
}
