% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NeighborhoodEstimation.R
\name{setupStabilityIndex}
\alias{setupStabilityIndex}
\title{Set up the subsamples that will be taken for stability selection}
\usage{
setupStabilityIndex(obs, strata = rep(1, nrow(obs)), B = 50, seed = 12345)
}
\arguments{
\item{obs}{a matrix of observations from which rows will be sampled}

\item{strata}{an optional vector of stratifying covariates, over which the sampling will be balanced}

\item{B}{Number of resamples}

\item{seed}{random seed}
}
\value{
list length \code{B} of indices
}
\description{
Stability selection is samples without replacement.  This initializes the sampling indices so that we can parallelize without worrying about the state of the random number generator (and recover from errors)
}
