% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HurdleNormal-package.R
\docType{data}
\name{shalek2014_list}
\alias{shalek2014_list}
\title{Single Cell RNAseq Expression Data from Shalek (2014)}
\format{
An object of class \code{list} of length 3.
}
\usage{
shalek2014_list
}
\description{
The 1-hour LPS data for genes with at least 30% expression
after thresholding with \code{MAST} are reproduced here as a list containing:
}
\details{
\itemize{
\item exprs: Genes x Cells list of log+1 transformed TPM
\item cdata: \code{data.frame} of cellular covariates
\item fdata' \code{data.frame} of transcript covariates
}
}
\keyword{datasets}
