% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProcessNeighborhoods.R
\name{sparseCbind}
\alias{sparseCbind}
\title{Recursively flatten and bind together sparse matrices}
\usage{
sparseCbind(x, sparse = TRUE)
}
\arguments{
\item{x}{a matrix, or a list of matrices.}

\item{sparse}{should a sparse result be returned?}
}
\value{
a Matrix, each column being a flattened matrix.
}
\description{
Recursively flatten and bind together sparse matrices
}

