% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HurdleStructure.R
\name{unpackTheta}
\alias{unpackTheta}
\title{Convert from old style to new style parameter vector or vice versa}
\usage{
unpackTheta(theta, method = "unpack")
}
\arguments{
\item{theta}{parameter vector}

\item{unpack}{if TRUE convert from old style to new, else convert from new to old}
}
\value{
ordered parameter
}
\description{
Convert from old style to new style parameter vector or vice versa
}

