% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GroupLasso.R
\name{Block}
\alias{Block}
\title{Define parameter groups}
\usage{
Block(this.model, bidx, group = "components", penalty.scale = NULL, bset)
}
\arguments{
\item{this.model}{(optional) a matrix created by \code{\link{makeModel}}}

\item{bidx}{NOT IMPLEMENTED starting index of each group}

\item{group}{\code{character}: one of components, or none.}

\item{penalty.scale}{\code{numeric} of length bidx/bset}

\item{bset}{NOT IMPLEMENTED}
}
\value{
a list containing a data.table `map` giving the mapping between parameters, groups and penalty scales and some other components
}
\description{
Define parameter groups and relative penalization.
Only one of \code{this.model}, \code{bidx} or \code{bset} should be provided.
When \code{this.model} is provided, it is assumed to have been created by with \code{makeModel},
and the first column is assumed to be an unpenalized intercept, followed by pairs of interaction terms.
The grouping corresponds to the parameter set for a node.
Otherwise, only one of bidx or bset should be set, but this is not implemented yet.
}

