% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ZeroInflatedNetworks-penalized.R
\name{cgpaths}
\alias{cgpaths}
\title{Get a solution path for the CG model}
\usage{
cgpaths(theta, y.zif, this.model, lambda = seq(1, 0, by = 0.1),
  control = list(tol = 1e-06, maxrounds = 100, maxit = 1000, debug = 1))
}
\arguments{
\item{theta}{initial guess for parameter}

\item{y.zif}{(zero-inflated) response}

\item{this.model}{covariates}

\item{lambda}{penalty path desired}

\item{control}{optimization control parameters}
}
\value{
matrix of parameters, one row per lambda
}
\description{
Get a solution path for the CG model
}

