% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/GroupLasso.R
\name{cgpaths}
\alias{cgpaths}
\title{Get a solution path for the CG model}
\usage{
cgpaths(theta, y.zif, this.model, lambda = exp(seq(1, -5, length.out = 10)),
  control = list(tol = 1e-06, maxrounds = 300, maxit = 500, debug = 1))
}
\arguments{
\item{theta}{initial guess for parameter}

\item{y.zif}{(zero-inflated) response}

\item{this.model}{covariates}

\item{lambda}{penalty path desired}

\item{control}{optimization control parameters}
}
\value{
matrix of parameters, one row per lambda
}
\description{
Get a solution path for the CG model
}

