% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AutoLogistic.R
\name{autoLogistic}
\alias{autoLogistic}
\alias{fitHurdle}
\title{Fit the hurdle model coordinate-by-coordinate on a sample}
\usage{
autoLogistic(samp, fixed = NULL, nlambda = 200, lambda.min.ratio = 0.1,
  parallel = FALSE)

fitHurdle(samp, parallel = TRUE, checkpoint_dir = NULL,
  makeModelArgs = NULL, returnNodePaths = FALSE, indices, ...)
}
\arguments{
\item{samp}{matrix of data, columns are variables}

\item{fixed}{unpenalized covariates to include.  NOT IMPLEMENTED for autoLogistic yet.}

\item{nlambda}{number of knots in solution path}

\item{lambda.min.ratio}{minimum lambda as function lambda0, the smallest lambda such that all coordinates are zero}

\item{parallel}{parallelize over variables using "mclapply"?}

\item{checkpoint_dir}{(optional) directory to save the fit of each gene, useful for large problems.  If it exists, then completed genes will be automatically loaded.}

\item{makeModelArgs}{(optional) arguments passed to the model matrix function}

\item{returnNodePaths}{return node-wise output (solution paths and diagnostics for each node) as attribute `nodePaths`}

\item{indices}{(optional) subset of indices to fit, useful for cluster parallelization.  NOT IMPLEMENTED.}

\item{...}{passed to cgpaths}
}
\value{
list of fits, one per coordinate and an attribute "timing"
}
\description{
Fit the hurdle model coordinate-by-coordinate on a sample
}
\section{Functions}{
\itemize{
\item \code{autoLogistic}: Fit an auto-logistic (Ising) model to \code{samp} using glmnet
}}

