% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AutoLogistic.R
\name{autoLogistic}
\alias{autoLogistic}
\alias{fitHurdle}
\title{Fit the hurdle model coordinate-by-coordinate on a sample}
\usage{
autoLogistic(samp, fixed = NULL, nlambda = 200, lambda.min.ratio = 0.1)

fitHurdle(samp, nlambda = 100, control = list(tol = 0.001, maxrounds = 3000,
  debug = 0, stepsize = 1), lambda.min.ratio = 0.1, makeModelArgs = NULL,
  parallel = TRUE, checkpoint_dir = NULL, ...)
}
\arguments{
\item{samp}{matrix of data, columns are variables}

\item{nlambda}{How many points to fit in the path}

\item{lambda.min.ratio}{minimum lambda as a fraction of lambda0}

\item{control}{arguments passed to cgpaths}

\item{makeModelArgs}{arguments passed to the model matrix function}

\item{parallel}{parallelize over variables using "mclapply"?}

\item{checkpoint_dir}{(optional) directory to save the fit of each gene, useful for large problems.  If it exists, then completed genes will be automatically loaded.}

\item{...}{passed to cgpaths}
}
\value{
list of fits, one per coordinate and an attribute "timing"
}
\description{
Fit the hurdle model coordinate-by-coordinate on a sample
}
\section{Functions}{
\itemize{
\item \code{autoLogistic}: Fit an auto-logistic (Ising) model to \code{samp} using glmnet
}}

