% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NeighborhoodEstimation.R
\name{autoGLM}
\alias{autoGLM}
\alias{autoLogistic}
\alias{fitHurdle}
\title{Fit the hurdle model coordinate-by-coordinate on a sample}
\usage{
autoGLM(
  samp,
  fixed = NULL,
  parallel = FALSE,
  keepNodePaths = FALSE,
  checkpointDir = "ignored",
  nlambda = 200,
  lambda.min.ratio = 0.1,
  family = "binomial"
)

fitHurdle(
  samp,
  fixed = NULL,
  parallel = TRUE,
  keepNodePaths = FALSE,
  checkpointDir = NULL,
  makeModelArgs = NULL,
  indices,
  ...
)
}
\arguments{
\item{samp}{matrix of data, columns are variables}

\item{fixed}{data.frame of fixed covariates (to be conditioned upon)}

\item{parallel}{parallelize over variables using "mclapply"?}

\item{keepNodePaths}{return node-wise output (solution paths and diagnostics for each node) as attribute `nodePaths`}

\item{checkpointDir}{(optional) directory to save the fit of each node, useful for large problems.  If it exists, then completed nodes will be automatically loaded.}

\item{nlambda}{number of lambda values on grid (default 200).}

\item{lambda.min.ratio}{minimum lambda ratio (as a function of lambda0, where the first predictor enters; default .1)}

\item{family}{in the case of \code{autoGLM} one of "gaussian" or "logistic"}

\item{makeModelArgs}{(optional) arguments passed to the model matrix function}

\item{indices}{(optional) subset of indices to fit, useful for cluster parallelization.}

\item{...}{passed to cgpaths}
}
\value{
list of fits, one per coordinate and an attribute "timing"
}
\description{
Fit the hurdle model coordinate-by-coordinate on a sample
}
\section{Functions}{
\itemize{
\item \code{autoGLM}: Fit an auto-model (Ising or Gaussian) to \code{samp} using glmnet.  checkpointDir is currently ignored.
}}

\seealso{
neighborhoodToArray, autoGLM, interpolateEdges
}
