% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Geom3.R
\name{ggally_hurdle}
\alias{ggally_hurdle}
\title{Wrapper so ggally can use \code{geom_hurdle}}
\usage{
ggally_hurdle(data, mapping, ...)
}
\arguments{
\item{data}{}

\item{mapping}{}

\item{...}{}
}
\value{
plot
}
\description{
.. content for \details{} ..
}

