% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Geom3.R
\name{ggally_hurdle}
\alias{ggally_hurdle}
\title{Wrapper so ggally can use \code{geom_hurdle}}
\usage{
ggally_hurdle(data, mapping, lwd.regression = 1, lwd.axis = 2,
  size.point = 1, ...)
}
\arguments{
\item{data}{data in which plot is evaluated}

\item{mapping}{a ggplot \code{mapping}}

\item{lwd.regression}{line width for regression line}

\item{lwd.axis}{line width for conditional mean along axis}

\item{size.point}{size of points in scatter plot}

\item{...}{passed to stat_hurdle1d}
}
\value{
plot
}
\description{
Wrapper so ggally can use \code{geom_hurdle}
}
