% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProcessNeighborhoods.R
\name{interpolateEdges}
\alias{interpolateEdges}
\title{Interpolate set of adjacency matrix across a set of edges}
\usage{
interpolateEdges(array, knot, nknot = 100)
}
\arguments{
\item{array}{\code{neighborhoodToArray} output}

\item{knot}{optional vector of number of edges over which to conduct the interpolation.  If missing then we use something on the order of P*log(P), where P is the number of nodes.}

\item{nknot}{number knots over which to interpolate}
}
\value{
a list of `edgeInterp`: contains a list of length `nknot` of sparse adjacency matrices, `estEdges` the desired number of edges to be interpolated, `trueEdges` the actual number of edges, `BIC`, the interpolated BIC.
}
\description{
Interpolate set of adjacency matrix across a set of edges
}
