% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GroupLassoHurdle.R
\name{makeModel}
\alias{makeModel}
\title{Generate a design matrix corresponding to the 2-1-0 hurdle}
\usage{
makeModel(zif, nodeId, fixed = NULL, center = TRUE, scale = FALSE,
  conditionalCenter = TRUE, ..., tol = 0.001)
}
\arguments{
\item{zif}{zero-inflated covariates}

\item{nodeId}{names to be used for nodes.  `colnames` of `zif` will be used if provided.}

\item{fixed}{optional matrix of fixed predictors}

\item{center}{center design matrix}

\item{scale}{scale design matrix}

\item{conditionalCenter}{in positive predictors, center only the non-zero components
(design matrix will still be centered, and orthogonal to indicators)}

\item{...}{ignored}

\item{tol}{tolerance for declaring a parameter equal to zero}
}
\value{
design matrix, with attribute fixedCols giving indices of unpenalized intercept columns
}
\description{
Generate a design matrix corresponding to the 2-1-0 hurdle
}

