% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProcessNeighborhoods.R
\name{neighborhoodToArray}
\alias{neighborhoodToArray}
\title{Convert neighborhood estimates fit on differing lambda paths into adjacency matrices}
\usage{
neighborhoodToArray(pathList, nknots, vnames = NULL,
  summaryFun = summarySignedL1, nobs, self_edges = FALSE)
}
\arguments{
\item{pathList}{a list of paths (coefficients in columns, paths in rows)}

\item{nknots}{number of lambda knots.  Be default, use the number of lambda in each neighborhood in pathList}

\item{vnames}{a vector of names to be applied to the resulting adjacency matrix}

\item{vnames}{a vector of names to be applied to the resulting adjacency matrix. The \code{nodeId} in each element in the \code{pathList} will be used if omitted.}
}
\value{
list of (sparse) adjacency matrices, the number of non-zero elements (edges, before enforcing symmetry) for each matrix, the lambda for each matrix, the non-penalized, refitted pseudo log-likelihood, the number of parameters per edge, and the BIC
}
\description{
Join a series of paths fit node-wise into a list of adjacency matrices.
}
