% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HurdleStructure.R
\name{unpackTheta}
\alias{unpackTheta}
\title{Convert from old style to new style parameter vector or vice versa}
\usage{
unpackTheta(theta, method = "unpack")
}
\arguments{
\item{theta}{parameter vector}

\item{method}{\code{character} one of 'unpack', 'pack', or 'gradient'}
}
\value{
ordered parameter
}
\description{
This just used to bridge between the R and C++ likelihood calcs, hence is legacy code now.
}
